Fix compatibility with Pipeline plugin
Pipeline plugin doesn't call createBuildWrapper of
ValidatingStringParameterValue(s) created by this plugin. So, runtime
validation of the values is not performed.

Add early validation of user entered parameters value against regex in
ValidatingStringParameterDefinition::createValue(), so that build is
prevented also when using Pipeline plugin.

By doing this, if the user hit "Build" in the presence of some invalid
parameter, an error page is shown (catching Failure() exception) that
reports why build has failed.